requirements.txt
# ============================================================
# Automated Helmet Violation Detection & ALPR System
# Requirements — Optimized for Google Colab (T4) + local CPU
# ============================================================

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# Object Detection
ultralytics>=8.1.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Augmentation
albumentations>=1.3.0

# Dataset Management
roboflow>=1.1.0

# Tracking & Supervision
supervision>=0.17.0

# Evaluation & Metrics
scikit-learn>=1.3.0

# Utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
